Move menu to pages directory
