+++ src/app/elitefour/pages/menu/menu.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
-import {ExportDataModalComponent, ImportDataModalComponent} from '../modals';
-import {VERSION} from '../../../environments/version';
+import {ExportDataModalComponent, ImportDataModalComponent} from '../../modals';
+import {VERSION} from '../../../../environments/version';
 
 @Component({
   selector: 'app-menu',
